Update the method return type to "long"
We are now using "pluginRepository.count()" to find the plugins count, whose return type is long

+++ src/main/java/io/jenkins/pluginhealth/scoring/repository/PluginRepository.java
@@ -38,6 +38,7 @@ public interface PluginRepository extends JpaRepository<Plugin, Long> {
 
     @Query(
         value = "SELECT count(id) FROM plugins WHERE scm IS NOT NULL AND details -> ?1 ->> 'status' = ?2",
-        nativeQuery = true)
-    int getProbeRawData(String probeID, String status);
+        nativeQuery = true
+    )
+    long getProbeRawData(String probeID, String status);
 }
